requirements.txt
# Web scraping
playwright==1.42.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Data processing
pandas==2.2.1
pydantic==2.6.3

# Email generation
openai==1.13.3

# UI (optional)
streamlit==1.32.2

# Utilities
python-dotenv==1.0.1
tqdm==4.66.2
